moved route files to own folder. Fixed date format, and labels for textfields now line up properly
