+++ RESTroutes/postRoutes.js
@@ -11,7 +11,8 @@ const con = mysql.createConnection({
     host: process.env.DB_HOST,
     user: process.env.DB_USER,
     password: process.env.DB_PASS,
-    database: process.env.DB_NAME
+    database: process.env.DB_NAME,
+    dateStrings: true
 });
 
 const query = util.promisify(con.query).bind(con);
