+++ post.html
@@ -11,6 +11,11 @@
     <!-- Latest compiled JavaScript -->
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
 </head>
+<style>
+    label {
+        margin-bottom: 6px;
+    }
+</style>
 <body>
 <nav class="navbar navbar-default">
     <div class="container-fluid">
