+++ update.html
@@ -12,6 +12,11 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
     <script src="client_put.js"></script>
 </head>
+<style>
+    label {
+        margin-bottom: 6px;
+    }
+</style>
 <body>
 <nav class="navbar navbar-default">
     <div class="container-fluid">
@@ -64,7 +69,7 @@
                 <input type="text" id="mfr" name="mfr" disabled value=""><br>
                 <input type="text" id="model_id" name="model_id" disabled value=""><br>
                 <input type="text" id="model_name" name="model_name" value=""><br>
-                <input type="text" id="release_date" name="release_date" value=""><br>
+                <input type="date" id="release_date" name="release_date" value=""><br>
                 <input type="text" id="weight_g" name="weight_g" value=""><br>
                 <input type="text" id="display_size_inch" name="display_size_inch" value=""><br>
                 <input type="text" id="resolution" name="resolution" value=""><br>
